+++ .expeditor/scripts/release_habitat/build_component.sh
@@ -27,13 +27,12 @@ declare -g hab_binary
 install_release_channel_hab_binary "$BUILD_PKG_TARGET"
 import_keys
 
+hab pkg install chef/hab-studio -c acceptance
+
 echo "--- :zap: Cleaning up old studio, if present"
 ${hab_binary} studio rm
 
 echo "--- :habicat: Building components/${component} using ${hab_binary}"
-
-# Set the refresh channel to the release channel.
-export HAB_STUDIO_SECRET_HAB_REFRESH_CHANNEL="${channel}"
 
 HAB_BLDR_CHANNEL="${channel}" ${hab_binary} pkg build "components/${component}"
 source results/last_build.env