+++ components/hab/src/command/pkg/promote.rs
@@ -37,11 +37,19 @@ pub async fn start(ui: &mut UI,
                    channel: &ChannelIdent,
                    token: &str)
                    -> Result<()> {
+    use habitat_core::package::Identifiable;
+
     let api_client = Client::new(bldr_url, PRODUCT, VERSION, None)?;
 
     ui.begin(format!("Promoting {} ({}) to channel '{}'", ident, target, channel))?;
 
-    if channel != &ChannelIdent::stable() && channel != &ChannelIdent::unstable() {
+    let default_channel_for_ident = if ident.origin() == "core" {
+        channel == &ChannelIdent::base()
+    } else {
+        channel == &ChannelIdent::stable()
+    };
+
+    if !default_channel_for_ident && channel != &ChannelIdent::unstable() {
         match api_client.create_channel(&ident.hacky_get_origin(), channel, token)
                         .await
         {