+++ components/plan-build/bin/hab-plan-build-linux.sh
@@ -366,7 +366,7 @@ export HAB_BLDR_CHANNEL
 # has been set to something different.
 : "${HAB_FALLBACK_CHANNEL=stable}"
 # Use the refresh channel for dependencies in the core origin
-: "${HAB_REFRESH_CHANNEL:=stable}"
+: "${HAB_REFRESH_CHANNEL:=base}"
 # If we prefer to use local core deps then a locally installed
 # package in this origin will be used in preference to what is in the refresh
 : "${HAB_PREFER_LOCAL_CHEF_DEPS:=false}"
@@ -752,8 +752,10 @@ _install_dependency() {
     local channel="$HAB_BLDR_CHANNEL"
     if [[ -z "${NO_INSTALL_DEPS:-}" || ${NO_INSTALL_DEPS} == "false" ]]; then
         origin="$(echo "$dep" | cut -d "/" -f 1)"
-        if [[ $origin == "core" ]]; then
-            channel="$HAB_REFRESH_CHANNEL"
+        if [[ $origin == "core" || $origin == "chef" ]]; then
+            if [[ $origin == "core" ]]; then
+                channel="$HAB_REFRESH_CHANNEL"
+            fi
             if [[ $HAB_PREFER_LOCAL_CHEF_DEPS == "false" ]]; then
                 IGNORE_LOCAL="--ignore-local"
             fi