+++ CHANGELOG.md
@@ -1,16 +1,21 @@
 # Habitat CHANGELOG
 
-<!-- latest_release 2.0.166 -->
-## [2.0.166](https://github.com/habitat-sh/habitat/tree/2.0.166) (2025-07-16)
+<!-- latest_release 2.0.171 -->
+## [2.0.171](https://github.com/habitat-sh/habitat/tree/2.0.171) (2025-07-21)
 
 #### Merged Pull Requests
-- Upgrade hab license subcommand to clapv4 [#9763](https://github.com/habitat-sh/habitat/pull/9763) ([dikshagupta1](https://github.com/dikshagupta1))
+- Updated `hab ring` to use clap v4 [#9767](https://github.com/habitat-sh/habitat/pull/9767) ([sougata-progress](https://github.com/sougata-progress))
 <!-- latest_release -->
 
 <!-- release_rollup since=1.6.1215-->
 ### Changes since 1.6.1215 release
 
 #### Merged Pull Requests
+- Updated `hab ring` to use clap v4 [#9767](https://github.com/habitat-sh/habitat/pull/9767) ([sougata-progress](https://github.com/sougata-progress)) <!-- 2.0.171 -->
+- Using `base` channel as default channel for packages in `core` origin [#9717](https://github.com/habitat-sh/habitat/pull/9717) ([agadgil-progress](https://github.com/agadgil-progress)) <!-- 2.0.170 -->
+- do not brew install gpung on mac install tests [#9772](https://github.com/habitat-sh/habitat/pull/9772) ([mwrock](https://github.com/mwrock)) <!-- 2.0.169 -->
+- publish aarch64-linux docker image [#9771](https://github.com/habitat-sh/habitat/pull/9771) ([sajjaphani](https://github.com/sajjaphani)) <!-- 2.0.168 -->
+- fixes the release pipeline with chef origin [#9769](https://github.com/habitat-sh/habitat/pull/9769) ([mwrock](https://github.com/mwrock)) <!-- 2.0.167 -->
 - Upgrade hab license subcommand to clapv4 [#9763](https://github.com/habitat-sh/habitat/pull/9763) ([dikshagupta1](https://github.com/dikshagupta1)) <!-- 2.0.166 -->
 - CLI Usage fixes for Clap v4 [#9766](https://github.com/habitat-sh/habitat/pull/9766) ([agadgil-progress](https://github.com/agadgil-progress)) <!-- 2.0.165 -->
 - documentation updates to reflect move to chef origin [#9749](https://github.com/habitat-sh/habitat/pull/9749) ([mwrock](https://github.com/mwrock)) <!-- 2.0.164 -->